Split out the command line parsing from the main entry point. For now we can put this in the test/ root
